test-py/requirements.txt
python-dotenv==1.0.0
openai>=1.0.0
pydantic>=2.0.0
aiohttp>=3.8.0
pillow>=10.0.0
